sitemoguena/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==43.0.1
decorator==5.1.1
defusedxml==0.8.0rc2
Django==5.0.6
django-extensions==3.2.3
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
executing==2.0.1
hiredis==3.0.0
idna==3.10
ipython==8.26.0
jedi==0.19.1
matplotlib-inline==0.1.7
oauthlib==3.2.2
parso==0.8.4
pillow==11.0.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.10
pure-eval==0.2.2
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
python-social-auth==0.3.6
python3-openid==3.2.0
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
wcwidth==0.2.13
